Moved contains_cycle logic from Sierra-gen to lowering
