+++ crates/cairo-lang-lowering/src/graph_algorithms/cycles.rs
@@ -1,0 +1,28 @@
+use cairo_lang_diagnostics::Maybe;
+use cairo_lang_semantic::ConcreteFunctionWithBodyId;
+
+use crate::db::LoweringGroup;
+
+/// Query implementation of [LoweringGroup::contains_cycle].
+pub fn contains_cycle(
+    db: &dyn LoweringGroup,
+    function_id: ConcreteFunctionWithBodyId,
+) -> Maybe<bool> {
+    let direct_callees = db.concrete_function_with_body_direct_callees_with_body(function_id)?;
+    for callee in direct_callees {
+        if db.contains_cycle(callee)? {
+            return Ok(true);
+        }
+    }
+
+    Ok(false)
+}
+
+/// Cycle handling for [LoweringGroup::contains_cycle].
+pub fn contains_cycle_handle_cycle(
+    _db: &dyn LoweringGroup,
+    _cycle: &[String],
+    _function_id: &ConcreteFunctionWithBodyId,
+) -> Maybe<bool> {
+    Ok(true)
+}
