+++ crates/cairo-lang-lowering/src/graph_algorithms/mod.rs
@@ -1,3 +1,4 @@
 pub mod concrete_function_node;
+pub mod cycles;
 pub mod feedback_set;
 pub mod strongly_connected_components;
